requirements.txt
certifi==2019.3.9
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.3
mkl-fft==1.0.10
mkl-random==1.0.2
numpy==1.16.2
olefile==0.46
pandas==0.24.2
Pillow==5.4.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tornado==6.0.2
wincertstore==0.2
torch
torchvision
